Split contact form into client component
